requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
google.generativeai
fastmcp